Rename qmlmodel module to instrument_model
It only contains the one class, and due to its length is unlikely to have more classes added, so a more accurate name seems appropriate.
