+++ geometry_constructor/instrument_model.py
