+++ geometry_constructor/writers.py
@@ -2,7 +2,7 @@ import h5py
 from pprint import pprint
 from geometry_constructor.datamodel import Sample, Detector, PixelGrid, PixelMapping, CountDirection, Corner, \
     Geometry, OFFGeometry, CylindricalGeometry, Component
-from geometry_constructor.qmlmodel import InstrumentModel
+from geometry_constructor.instrument_model import InstrumentModel
 from PySide2.QtCore import QObject, QUrl, Slot
